Move file operations to a separate command file
