+++ blaze_explorer_lib/Cargo.toml
@@ -16,6 +16,7 @@ lazy_static = "1.5.0"
 libloading = "0.8.6"
 mockall = "0.13.0"
 open = "5.3.0"
+rand = "0.9.0"
 ratatui = "0.27.0"
 rust_search = "2.1.0"
 serde = {version="1.0.210", features = ["derive"]}
